FossilOrigin-Name: c4fca467f727745c95c61bfc2e725e2ddce3a433

+++ src/xfer.c
@@ -96,7 +96,7 @@ static void remote_has(int rid){
 }
 
 /*
-** The aToken[0..nToken-1] blob array is a parse of a "file" line 
+** The aToken[0..nToken-1] blob array is a parse of a "file" line
 ** message.  This routine finishes parsing that message and does
 ** a record insert of the file.
 **
@@ -121,10 +121,10 @@ static void xfer_accept_file(Xfer *pXfer, int cloneFlag){
   int srcid = 0;
   Blob content, hash;
   int isPriv;
-  
+
   isPriv = pXfer->nextIsPrivate;
   pXfer->nextIsPrivate = 0;
-  if( pXfer->nToken<3 
+  if( pXfer->nToken<3
    || pXfer->nToken>4
    || !blob_is_uuid(&pXfer->aToken[1])
    || !blob_is_int(&pXfer->aToken[pXfer->nToken-1], &n)
@@ -200,7 +200,7 @@ static void xfer_accept_file(Xfer *pXfer, int cloneFlag){
 }
 
 /*
-** The aToken[0..nToken-1] blob array is a parse of a "cfile" line 
+** The aToken[0..nToken-1] blob array is a parse of a "cfile" line
 ** message.  This routine finishes parsing that message and does
 ** a record insert of the file.  The difference between "file" and
 ** "cfile" is that with "cfile" the content is already compressed.
@@ -229,10 +229,10 @@ static void xfer_accept_compressed_file(Xfer *pXfer){
   int srcid = 0;
   Blob content;
   int isPriv;
-  
+
   isPriv = pXfer->nextIsPrivate;
   pXfer->nextIsPrivate = 0;
-  if( pXfer->nToken<4 
+  if( pXfer->nToken<4
    || pXfer->nToken>5
    || !blob_is_uuid(&pXfer->aToken[1])
    || !blob_is_int(&pXfer->aToken[pXfer->nToken-2], &szU)
@@ -286,7 +286,7 @@ static int send_delta_parent(
     "SELECT pid FROM plink x"
     " WHERE cid=%d"
     "   AND NOT EXISTS(SELECT 1 FROM phantom WHERE rid=pid)",
-    
+
     "SELECT pid, min(mtime) FROM mlink, event ON mlink.mid=event.objid"
     " WHERE fid=%d"
     "   AND NOT EXISTS(SELECT 1 FROM phantom WHERE rid=pid)"
@@ -323,7 +323,7 @@ static int send_delta_parent(
 }
 
 /*
-** Try to send a file as a native delta.  
+** Try to send a file as a native delta.
 ** If successful, return the number of bytes in the delta.
 ** If we cannot generate an appropriate delta, then send
 ** nothing and return zero.
@@ -405,7 +405,7 @@ static void send_file(Xfer *pXfer, int rid, Blob *pUuid, int nativeDelta){
     blob_reset(&uuid);
     return;
   }
-  if( (pXfer->maxTime != -1 && time(NULL) >= pXfer->maxTime) || 
+  if( (pXfer->maxTime != -1 && time(NULL) >= pXfer->maxTime) ||
        pXfer->mxSend<=blob_size(pXfer->pOut) ){
     const char *zFormat = isPriv ? "igot %b 1\n" : "igot %b\n";
     blob_appendf(pXfer->pOut, zFormat, pUuid);
@@ -447,7 +447,7 @@ static void send_file(Xfer *pXfer, int rid, Blob *pUuid, int nativeDelta){
 
 /*
 ** Send the file identified by rid as a compressed artifact.  Basically,
-** send the content exactly as it appears in the BLOB table using 
+** send the content exactly as it appears in the BLOB table using
 ** a "cfile" card.
 */
 static void send_compressed_file(Xfer *pXfer, int rid){
@@ -517,7 +517,7 @@ static void send_compressed_file(Xfer *pXfer, int rid){
 */
 static void request_phantoms(Xfer *pXfer, int maxReq){
   Stmt q;
-  db_prepare(&q, 
+  db_prepare(&q,
     "SELECT uuid FROM phantom JOIN blob USING(rid)"
     " WHERE NOT EXISTS(SELECT 1 FROM shun WHERE uuid=blob.uuid) %s",
     (pXfer->syncPrivate ? "" :
@@ -553,8 +553,8 @@ static int check_tail_hash(Blob *pHash, Blob *pMsg){
 **
 **        login LOGIN NONCE SIGNATURE
 **
-** The NONCE is the SHA1 hash of the remainder of the input.  
-** SIGNATURE is the SHA1 checksum of the NONCE concatenated 
+** The NONCE is the SHA1 hash of the remainder of the input.
+** SIGNATURE is the SHA1 checksum of the NONCE concatenated
 ** with the users password.
 **
 ** The parameters to this routine are ephemeral blobs holding the
@@ -566,7 +566,7 @@ static int check_tail_hash(Blob *pHash, Blob *pMsg){
 **
 ** If anything fails to check out, no changes are made to privileges.
 **
-** Signature generation on the client side is handled by the 
+** Signature generation on the client side is handled by the
 ** http_exchange() routine.
 **
 ** Return non-zero for a login failure and zero for success.
@@ -580,7 +580,8 @@ int check_login(Blob *pLogin, Blob *pNonce, Blob *pSig){
   if( fossil_strcmp(zLogin, "nobody")==0 || fossil_strcmp(zLogin,"anonymous")==0 ){
     return 0;   /* Anybody is allowed to sync as "nobody" or "anonymous" */
   }
-  if( fossil_strcmp(P("REMOTE_USER"), zLogin)==0 ){
+  if( fossil_strcmp(P("REMOTE_USER"), zLogin)==0
+      && db_get_boolean("remote_user_ok",0) ){
     return 0;   /* Accept Basic Authorization */
   }
   db_prepare(&q,
@@ -701,7 +702,7 @@ void create_cluster(void){
     }
     db_finalize(&q);
     db_multi_exec(
-      "DELETE FROM unclustered WHERE rid NOT IN (0 %s)", 
+      "DELETE FROM unclustered WHERE rid NOT IN (0 %s)",
       blob_str(&deleteWhere)
     );
     blob_reset(&deleteWhere);
@@ -740,7 +741,7 @@ static int send_unclustered(Xfer *pXfer){
   Stmt q;
   int cnt = 0;
   if( pXfer->resync ){
-    db_prepare(&q, 
+    db_prepare(&q,
       "SELECT uuid, rid FROM blob"
       " WHERE NOT EXISTS(SELECT 1 FROM shun WHERE uuid=blob.uuid)"
       "   AND NOT EXISTS(SELECT 1 FROM phantom WHERE rid=blob.rid)"
@@ -750,7 +751,7 @@ static int send_unclustered(Xfer *pXfer){
       pXfer->resync
     );
   }else{
-    db_prepare(&q, 
+    db_prepare(&q,
       "SELECT uuid FROM unclustered JOIN blob USING(rid)"
       " WHERE NOT EXISTS(SELECT 1 FROM shun WHERE uuid=blob.uuid)"
       "   AND NOT EXISTS(SELECT 1 FROM phantom WHERE rid=blob.rid)"
@@ -774,7 +775,7 @@ static int send_unclustered(Xfer *pXfer){
 */
 static void send_all(Xfer *pXfer){
   Stmt q;
-  db_prepare(&q, 
+  db_prepare(&q,
     "SELECT uuid FROM blob "
     " WHERE NOT EXISTS(SELECT 1 FROM shun WHERE uuid=blob.uuid)"
     "   AND NOT EXISTS(SELECT 1 FROM private WHERE rid=blob.rid)"
@@ -1035,8 +1036,8 @@ void page_xfer(void){
         }
       }
     }else
-  
-    
+
+
     /*    pull  SERVERCODE  PROJECTCODE
     **    push  SERVERCODE  PROJECTCODE
     **
@@ -1144,10 +1145,10 @@ void page_xfer(void){
           @ error login\sfailed
           nErr++;
           break;
-        }        
+        }
       }
     }else
-    
+
     /*    reqconfig  NAME
     **
     ** Request a configuration value
@@ -1169,7 +1170,7 @@ void page_xfer(void){
         }
       }
     }else
-    
+
     /*   config NAME SIZE \n CONTENT
     **
     ** Receive a configuration value from the client.  This is only
@@ -1196,7 +1197,7 @@ void page_xfer(void){
       blob_seek(xfer.pIn, 1, BLOB_SEEK_CUR);
     }else
 
-      
+
 
     /*    cookie TEXT
     **
@@ -1376,7 +1377,7 @@ static double fossil_fabs(double x){
 }
 
 /*
-** Sync to the host identified in g.urlName and g.urlPath.  This
+** Sync to the host identified in g.url.name and g.url.path.  This
 ** routine is called by the client.
 **
 ** Records are pushed to the server if pushFlag is true.  Records
@@ -1415,7 +1416,7 @@ int client_sync(
   double rSkew = 0.0;     /* Maximum time skew */
 
   if( db_get_boolean("dont-push", 0) ) syncFlags &= ~SYNC_PUSH;
-  if( (syncFlags & (SYNC_PUSH|SYNC_PULL|SYNC_CLONE))==0 
+  if( (syncFlags & (SYNC_PUSH|SYNC_PULL|SYNC_CLONE))==0
      && configRcvMask==0 && configSendMask==0 ) return 0;
 
   transport_stats(0, 0, 1);
@@ -1430,11 +1431,6 @@ int client_sync(
     xfer.syncPrivate = 1;
   }
 
-  db_begin_transaction();
-  db_record_repository_filename(0);
-  db_multi_exec(
-    "CREATE TEMP TABLE onremote(rid INTEGER PRIMARY KEY);"
-  );
   blobarray_zero(xfer.aToken, count(xfer.aToken));
   blob_zero(&send);
   blob_zero(&recv);
@@ -1473,7 +1469,6 @@ int client_sync(
     if( (syncFlags & SYNC_PULL)==0 ) zOpType = "Push";
     if( (syncFlags & SYNC_RESYNC)!=0 ) xfer.resync = 0x7fffffff;
   }
-  manifest_crosslink_begin();
   if( syncFlags & SYNC_VERBOSE ){
     fossil_print(zLabelFormat, "", "Bytes", "Cards", "Artifacts", "Deltas");
   }
@@ -1481,6 +1476,12 @@ int client_sync(
   while( go ){
     int newPhantom = 0;
     char *zRandomness;
+    db_begin_transaction();
+    db_record_repository_filename(0);
+    db_multi_exec(
+      "CREATE TEMP TABLE onremote(rid INTEGER PRIMARY KEY);"
+    );
+    manifest_crosslink_begin();
 
     /* Send make the most recently received cookie.  Let the server
     ** figure out if this is a cookie that it cares about.
@@ -1489,7 +1490,7 @@ int client_sync(
     if( zCookie ){
       blob_appendf(&send, "cookie %s\n", zCookie);
     }
-    
+
     /* Generate gimme cards for phantoms and leaf cards
     ** for all leaves.
     */
@@ -1505,7 +1506,7 @@ int client_sync(
     }
 
     /* Send configuration parameter requests.  On a clone, delay sending
-    ** this until the second cycle since the login card might fail on 
+    ** this until the second cycle since the login card might fail on
     ** the first cycle.
     */
     if( configRcvMask && ((syncFlags & SYNC_CLONE)==0 || nCycle>0) ){
@@ -1560,6 +1561,7 @@ int client_sync(
     if( http_exchange(&send, &recv, (syncFlags & SYNC_CLONE)==0 || nCycle>0,
         MAX_REDIRECTS) ){
       nErr++;
+      go = 2;
       break;
     }
 
@@ -1665,7 +1667,7 @@ int client_sync(
           if( rid ) send_file(&xfer, rid, &xfer.aToken[1], 0);
         }
       }else
-  
+
       /*   igot UUID  ?PRIVATEFLAG?
       **
       ** Server announces that it has a particular file.  If this is
@@ -1674,7 +1676,7 @@ int client_sync(
       ** Always remember that the server has this file so that we do
       ** not transmit it by accident.
       **
-      ** If the PRIVATE argument exists and is 1, then the file is 
+      ** If the PRIVATE argument exists and is 1, then the file is
       ** private.  Pretend it does not exists if we are not pulling
       ** private files.
       */
@@ -1695,8 +1697,8 @@ int client_sync(
         }
         remote_has(rid);
       }else
-    
-      
+
+
       /*   push  SERVERCODE  PRODUCTCODE
       **
       ** Should only happen in response to a clone.  This message tells
@@ -1718,7 +1720,7 @@ int client_sync(
         if( cloneSeqno>0 ) blob_appendf(&send, "clone 3 %d\n", cloneSeqno);
         nCardSent++;
       }else
-      
+
       /*   config NAME SIZE \n CONTENT
       **
       ** Receive a configuration value from the server.
@@ -1740,7 +1742,7 @@ int client_sync(
         blob_seek(xfer.pIn, 1, BLOB_SEEK_CUR);
       }else
 
-      
+
       /*    cookie TEXT
       **
       ** The server might include a cookie in its reply.  The client
@@ -1782,7 +1784,7 @@ int client_sync(
       **
       ** If the "login failed" message is seen, clear the sync password prior
       ** to the next cycle.
-      */        
+      */
       if( blob_eq(&xfer.aToken[0],"message") && xfer.nToken==2 ){
         char *zMsg = blob_terminate(&xfer.aToken[1]);
         defossilize(zMsg);
@@ -1799,7 +1801,7 @@ int client_sync(
       /*    pragma NAME VALUE...
       **
       ** The server can send pragmas to try to convey meta-information to
-      ** the client.  These are informational only.  Unknown pragmas are 
+      ** the client.  These are informational only.  Unknown pragmas are
       ** silently ignored.
       */
       if( blob_eq(&xfer.aToken[0], "pragma") && xfer.nToken>=2 ){
@@ -1812,10 +1814,10 @@ int client_sync(
       ** Except, when cloning we will sometimes get an error on the
       ** first message exchange because the project-code is unknown
       ** and so the login card on the request was invalid.  The project-code
-      ** is returned in the reply before the error card, so second and 
+      ** is returned in the reply before the error card, so second and
       ** subsequent messages should be OK.  Nevertheless, we need to ignore
       ** the error card on the first message of a clone.
-      */        
+      */
       if( blob_eq(&xfer.aToken[0],"error") && xfer.nToken==2 ){
         if( (syncFlags & SYNC_CLONE)==0 || nCycle>0 ){
           char *zMsg = blob_terminate(&xfer.aToken[1]);
@@ -1824,11 +1826,11 @@ int client_sync(
           fossil_print("Error: %s\n", zMsg);
           if( fossil_strcmp(zMsg, "login failed")==0 ){
             if( nCycle<2 ){
-              g.urlPasswd = 0;
+              g.url.passwd = 0;
               go = 1;
               if( g.cgiOutput==0 ){
-                g.urlFlags |= URL_PROMPT_PW;
-                g.urlFlags &= ~URL_PROMPTED;
+                g.url.flags |= URL_PROMPT_PW;
+                g.url.flags &= ~URL_PROMPTED;
                 url_prompt_for_password();
                 url_remember();
               }
@@ -1896,7 +1898,7 @@ int client_sync(
     xfer.nDanglingFile = 0;
 
     /* If we have one or more files queued to send, then go
-    ** another round 
+    ** another round
     */
     if( xfer.nFileSent+xfer.nDeltaSent>0 ){
       go = 1;
@@ -1910,7 +1912,15 @@ int client_sync(
     ** on a clone in order to be sure to retrieve the configuration
     ** information which is only sent on the second round.
     */
-    if( cloneSeqno<=0 && nCycle>1 ) go = 0;   
+    if( cloneSeqno<=0 && nCycle>1 ) go = 0;
+    db_multi_exec("DROP TABLE onremote");
+    if( go ){
+      manifest_crosslink_end(MC_PERMIT_HOOKS);
+    }else{
+      manifest_crosslink_end(MC_PERMIT_HOOKS);
+      content_enable_dephantomize(1);
+    }
+    db_end_transaction(0);
   };
   transport_stats(&nSent, &nRcvd, 1);
   if( (rSkew*24.0*3600.0) > 10.0 ){
@@ -1927,11 +1937,13 @@ int client_sync(
   fossil_print(
      "%s finished with %lld bytes sent, %lld bytes received\n",
      zOpType, nSent, nRcvd);
-  transport_close(GLOBAL_URL());
-  transport_global_shutdown(GLOBAL_URL());
-  db_multi_exec("DROP TABLE onremote");
-  manifest_crosslink_end(MC_PERMIT_HOOKS);
-  content_enable_dephantomize(1);
-  db_end_transaction(0);
+  transport_close(&g.url);
+  transport_global_shutdown(&g.url);
+  if( nErr && go==2 ){
+    db_multi_exec("DROP TABLE onremote");
+    manifest_crosslink_end(MC_PERMIT_HOOKS);
+    content_enable_dephantomize(1);
+    db_end_transaction(0);
+  }
   return nErr;
 }
